refactor(lsp): split signature help helpers
