+++ scripts/arch/arch_guard_policy.toml
@@ -303,7 +303,7 @@ exclude_files = [
   "crates/tsz-checker/src/types/class_type/core.rs",
   "crates/tsz-emitter/src/declaration_emitter/helpers/mod.rs",
   "crates/tsz-emitter/src/declaration_emitter/helpers/type_printing.rs",
-  "crates/tsz-lsp/src/signature_help.rs",
+  "crates/tsz-lsp/src/signature_help/shapes.rs",
 ]
 ignore_comment_lines = true
 
@@ -335,7 +335,7 @@ exclude_files = [
   # file_id_allocator.lookup() is not a solver interner lookup
   "crates/tsz-lsp/src/project/core.rs",
   # Pre-existing baseline debt
-  "crates/tsz-lsp/src/signature_help.rs",
+  "crates/tsz-lsp/src/signature_help/shapes.rs",
 ]
 
 [[pattern_checks]]
